imade/appengine/building-an-app/requirements.txt
Flask==2.0.2
google-cloud-datastore==2.4.0
google-auth==2.5.0
requests==2.27.1